Add wrap-around movement to the main play board...
This introduces an interesting bug, where the player can travel under the wall and appear on the other side...

This is because we teleport to the spot under the wall, so that the collision can be detected correctly on the next movement. This is important when the food spawns right next to the wall.

I will fix this (somehow), but want to tag this point for posterity.